requirements.txt
yfinance==0.2.66
rich==14.3.1
python-dotenv==1.1.1
